docs: Cloudflare R2 Bucket storage integration
I rewrote the instructions for creating an R2 Bucket on Cloudflare because they were incorrect/outdated. The procedure for requesting Access Keys has changed, they are now known as Tokens. I've updated the instructions to reflect the new menu points. The configuration of the R2 service in the Storage Objects on MEMOS referred to a menu not present in the latest release; I updated the definitions and corrected the procedure for activating public viewing. It should be noted that public viewing is probably in "legacy mode" on Cloudflare and may perhaps disappear in future versions of R2 Bucket, in favor of the use of workers instead. (more info here: https://developers.cloudflare.com/r2/buckets/public-buckets/ )
This pull request completely replaces the previous one.